Revert "Minor Update."
This reverts commit 78effcd1fb52ffae8085f119253fcf942053563f.

+++ sku.0/sys.server/compiled/game/script/library/reverse_engineering.java
@@ -40,7 +40,7 @@ public class reverse_engineering extends script.base_script
     public static final String JUNK_TABLE = "datatables/crafting/reverse_engineering_junk.iff";
     public static final string_id POWERUP_PID_TITLE = new string_id("spam", "powerup_override_title");
     public static final string_id POWERUP_PID_PROMPT = new string_id("spam", "powerup_override");
-    public static final int EXPIRATION_TIME = utils.getIntConfigSetting("GameServer", "powerupBuffDuration");
+    public static final float EXPIRATION_TIME = utils.getIntConfigSetting("GameServer", "powerupBuffDuration");
     //public static final float EXPIRATION_TIME = 1800.0f;
     
     public static void checkPowerUpReApply(obj_id player) throws InterruptedException
